refeita a lógica de renderizar os matches em outra página, agora tudo está funcionando
