+++ semana9/semana9-projeto/astromatch/astromatch/src/components/Home.js
@@ -7,15 +7,14 @@ import {
   OptionsIcon,
   OptionsButton,
   ShowListButton,
-  ClearButton
 } from "./styled-components";
 import IconLike from "../img/like-pink.png";
 import IconRemove from "../img/remove.jpg";
 
 export default function Home() {
   const [showPerfile, setShowPerfile] = useState([]);
-  const [showPage, setShowPage] = useState(true);
-  const [isMatch, setIsMatch] = useState("");
+  const [showPage, setShowPage] = useState('');
+//   const [isMatch, setIsMatch] = useState("");
   const [answer, setAnswer] = useState(true);
 
   useEffect(() => {
@@ -47,7 +46,7 @@ export default function Home() {
     axios
       .post(`${baseUrl}/${axiosConfig}/choose-person`, body)
       .then((response) => {
-        setIsMatch(response.data.isMatch);
+        // setIsMatch(response.data.isMatch);
         setAnswer(true);
         alert("Uhuuu, você curtiu!");
       })
@@ -64,7 +63,6 @@ export default function Home() {
     axios
       .post(`${baseUrl}/${axiosConfig}/choose-person`, body)
       .then((response) => {
-        setIsMatch(response.data.isMatch);
         setAnswer(false);
         alert("Que pena, você não curtiu :(");
       })
@@ -73,14 +71,11 @@ export default function Home() {
       });
   };
 
-  const clearMatchs = () => {
-      axios.put(`${baseUrl}/${axiosConfig}/clear`)
-      .then((response) => {
-        alert('Seus matches foram removidos com sucesso!')
-      })
-      .catch((error) => {
-          console.log(error)
-      })
+  
+  if(showPage) {
+      return <Matches />
+  } else if (showPage !==showPage) {
+      return <Home />
   }
 
   return (
@@ -106,9 +101,7 @@ export default function Home() {
           />
         </OptionsButton>
         <div>
-          {showPage ? null : <Matches />}
           <ShowListButton onClick={changePage}>Mostrar Matches</ShowListButton>
-          <ClearButton onClick={() => {if(window.confirm("Tem certeza que deseja limpar a sua lista de matches?")) {clearMatchs(showPerfile.id)}}}>Limpar Matches</ClearButton>
         </div>
       </div>
     </div>
